requirements.txt
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2-binary==2.9.6
requests==2.28.1
prettytable==3.8.0